Add files via upload
i'm pretty sure that this works and creates a correct waitlist and enrollment list using pythons "sorted" function. but the issue is that it somehow deletes the information in my dictionary, even though when i created the function i make a copy of the dictionary and then work with the copy.